labs/python/requirements.txt
agent-framework-azure-ai==1.0.0rc4
agent-framework-core==1.0.0rc4
python-dotenv
azure-identity
azure-ai-projects
azure-search-documents
azure-mgmt-cognitiveservices
azure-core

# MCP Dependencies
mcp
httpx
httpx-sse

# Azure OpenAI
openai

# REST API Server
fastapi
uvicorn
pydantic
starlette

# Async support
anyio